+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - This works back to Node 16.19.0
 - Use unreleased fixed version of node-getopt
 - On server start error output error to console
+- URL option to not minify
 
 
 ## 8.0.0 : 10 Feb 2023